added and removed code from current routes
