+++ routes/aggregatedstats.js
@@ -1,23 +1,17 @@
 const express = require('express');
 const router = express.Router();
-const axios = require('axios');
 const mysql = require('mysql');
 const chalk = require('chalk');
-const utf8 = require('utf8');
 const moment = require('moment');
 const _ = require('underscore');
 const winston = require('winston');
 
 let timestampForRequest;        // can be made const
 
-const serverIp = process.env.SERVERIP || (() => { new Error("Provide a server IP in env vars") });
-const endpoint = process.env.APIENDPOINT || (() => { new Error("Provide a api endpoint in env vars") });
 const dbHost = process.env.DBENDPOINT || (() => { new Error("Provide a db endpoint in env vars") });
 const dbPassword = process.env.DBPASSWORD || (() => { new Error("Provide a db password in env vars") });
 const dbUsername = process.env.DBUSER || (() => { new Error("Provide a db username in env vars") });
 const dbName = process.env.DBNAME || (() => { new Error("Provide a db username in env vars") });
-
-const keyword = keyword => chalk.keyword('blue')(keyword)
 
 const dir = './logging/'
 
@@ -57,9 +51,9 @@ router.get('/', async (req, res) => {
 
 router.get('/playerCount', async (req, res) => {
     const duration = req.query.duration ? req.query.duration : 288
-pool.getConnection((err, connection) => {
-            if (err) console.log(err);
-    connection.query(`SELECT
+    pool.getConnection((err, connection) => {
+        if (err) console.log(err);
+        connection.query(`SELECT
     time, playerCount
     FROM
         (
@@ -67,18 +61,18 @@ pool.getConnection((err, connection) => {
         ) a
     ORDER BY
     time;`, (err, result, fields) => {
-                if (err) console.log(err);
-                if (result) {
-                    res.status(200).json({
-                        duration,
-                        response: result
-                    })
-                    console.log(chalk.blue('Completed query for ' + chalk.whiteBright.underline(duration) + " records at aggregatedstats/playerCount"))
-                }
-            });
-    connection.release();
-    if (err) throw err;
+            if (err) console.log(err);
+            if (result) {
+                res.status(200).json({
+                    duration,
+                    response: result
+                })
+                console.log(chalk.blue('Completed query for ' + chalk.whiteBright.underline(duration) + " records at aggregatedstats/playerCount"))
+            }
         });
+        connection.release();
+        if (err) throw err;
+    });
 })
 
 // TODO: Do we need post for playerCount?
