+++ routes/serverstats.js
@@ -21,6 +21,7 @@ const pointsSpentUtil = require('../serverStatsUtils/pointsSpent')
 const resetDailyUtil = require('../serverStatsUtils/resetDaily')
 const serverInfoUtil = require('../serverStatsUtils/serverInfo')
 const serverStatsUtil = require('../serverStatsUtils/serverStats')
+const temporaryDataUtil = require('../serverStatsUtils/temporaryData')
 
 let directQueryInfo = {};
 let directPlayerInfo = [];
@@ -41,7 +42,6 @@ const dbName = process.env.DBNAME || (() => { new Error("Provide a db username i
 const basicAuthUsername = process.env.BASICAUTHUSERNAME || (() => { new Error("Provide a server IP in env vars") });
 const basicAuthPassword = process.env.BASICAUTHPASSWORD || (() => { new Error("Provide a server IP in env vars") });
 
-const recognisedTableNames = ['aggregatedInfo', 'playerInfo', 'playersComparisonFirst', 'playersComparisonSecond', 'serverInfo']
 const recognisedTemporaryTableNames = ['playersComparisonFirst', 'playersComparisonSecond']
 
 const users = {};
@@ -165,334 +165,16 @@ router.post('/pointsSpent', async (req, res) => {
         })
 })
 
-router.get('/allRows', async (req, res) => {
-    if (req.query.tableName && recognisedTableNames.includes(req.query.tableName)) {
-        pool.getConnection((err, connection) => {
-            if (err) console.error(err)
-            connection.query(`SELECT * FROM ${req.query.tableName}`, (error, rows, fields) => {
-                if (error) console.log(error);
-                res.status(200).json({
-                    message: `Successfully got all data from ${req.query.tableName}`,
-                    result: rows
-                });
-            });
-            connection.release();
-            if (err) throw err;
-        });
-    }
-    else {
-        res.status(400).json({
-            error: {
-                message: `Please provide valid table name`,
-            }
-        });
-        console.log('Not valid table name');
-    }
-});
-
-router.get('/resetDaily', async (req, res) => {
-    pool.getConnection((err, connection) => {
-        if (err) console.error(err)
-        pool.query(`UPDATE playerInfo SET totalKillsDaily = 0, totalPointsSpentDaily = 0, totalTimeDaily = 0 WHERE playerName IS NOT NULL`,
-            (err, result, fields) => {
-                if (err) console.error(err)
-                res.status(200).json({
-                    message: "Reset totalKillsDaily, totalPointsSpentDaily, totalTimeDaily to 0"
-                })
-            });
-        console.log(chalk.blue('Reset totalKillsDaily, totalPointsSpentDaily, totalTimeDaily to ' + chalk.whiteBright.underline(keyword('0'))))
-        connection.release();
-        if (err) throw err;
-    });
-})
-
 router.post('/temporaryData', async (req, res) => {
-    if (req.query.name && req.query.time && req.query.score && req.query.tableName && recognisedTemporaryTableNames.includes(req.query.tableName)) {
-        pool.getConnection((err, connection) => {
-            const tableName = req.query.tableName;
-            const name = utf8decode(decodeURIComponent(req.query.name));
-            const time = req.query.time;
-            const score = req.query.score;
-
-            connection.query(`INSERT INTO ${tableName} (name, time, score)
-            VALUES (${sqlString.escape(name)}, ${time}, ${score}) ON DUPLICATE KEY UPDATE name = ${sqlString.escape(name)}, time = ${time}, score = ${score}`, (err, result, fields) => {
-                if (err) console.log(err);
-                if (result) {
-                    res.status(201).json({
-                        message: `Created temporary player inside ${req.query.tableName}`
-                    })
-                    console.log(chalk.blue('Database entry ' + chalk.whiteBright.underline(keyword(name)) + " with duration " + chalk.whiteBright.underline(keyword(time)) + " and score " + chalk.whiteBright.underline(keyword(score)) + ' added into ' + req.query.tableName))
-                }
-                if (fields) console.log(fields);
-                connection.release();
-                if (err) throw err;
-            });
-        });
-    } else {
-        res.status(400).json({
-            error: {
-                message: 'Please provide name, time, score and a valid tableName in request',
-            }
-        });
-        console.log('Need to provide name, time, score and valid tableName in request');
-    }
-})
-
-router.get('/repeatedRequests', async (req, res) => {
-    if (!running) {
-        running = true;
-
-        timestamp = moment().format('YYYY-MM-DD HH:mm:ss')
-
-        console.log("GET: repeatedRequests " + timestamp)
-
-        res.status(200).json({
-            message: "Initiating repeated requests ", timestamp
+    temporaryDataUtil(req.query.name, req.query.time, req.query.score, req.query.tableName, pool, recognisedTemporaryTableNames)
+        .then(result => {
+            res.status(201).json({ message: `Created temporary player inside ${result.tableName}`})
+            console.log(chalk.blue('Database entry ' + chalk.whiteBright.underline(keyword(result.name)) + " with duration " + chalk.whiteBright.underline(keyword(result.time)) + " and score " + chalk.whiteBright.underline(keyword(result.score)) + ' added into ' + result.tableName))
         })
-
-        // sends player count every 5 mins to serverInfo table
-        cron.schedule('*/5 * * * *', async () => {
-            const serverInfo = await getNewPlayers()
-                .then(eachObject => (
-                    eachObject
-                        .map(element => element.directQueryInfo)
-                        .filter(el => el != null)))
-
-        const undefinedCheck = (objectToCheck, ifUndefined) => objectToCheck == undefined ? ifUndefined : objectToCheck;
-        serverInfoUtil(serverInfo[0].playersnum, serverInfo[0].botsnum, serverInfo[0].name, serverInfo[0].map, undefinedCheck, pool);
-
-        });
-
-        // sends query to set all daily columns to 0 at 00:01 everyday
-        cron.schedule('01 00 * * *', async () => {
-            resetDailyUtil();
+        .catch(result => {
+            console.log(chalk.red(result))
+            res.status(400).json({ message: result })
         })
-
-        // initial request of players every 15 seconds to playersComparisonCache table
-        const firstJob = schedule.scheduleJob({ rule: '*/15 * * * * *' }, async (fireDate) => {
-
-            const serverInfoUnfiltered = await getNewPlayers();
-            const serverInfo = serverInfoUnfiltered
-                .map(element => element.directQueryInfo)
-                .filter(el => el != null)
-
-            if (!serverInfo[0].name) {
-                console.log(chalk.magentaBright("Server not online! Waiting for 1 minute"))
-                firstJob.cancel(true)
-                secondJob.cancel(true)
-                firstJob.cancelNext(true)
-                secondJob.cancelNext(true)
-            }
-
-            console.log(chalk.hex('#DEADED').bold('running first task ************************************************************************************************************************************************************'));
-            console.log('First starting now: ' + moment().format('ss') + " but really started at " + fireDate);
-
-            const playersToBePushedToTemporaryTable = [];
-
-            const playerInfoToBeCompared = await getNewPlayers()
-                .then(eachObject => (
-                    eachObject
-                        .map(element => element.directPlayerInfo)
-                        .filter(el => el != null)))
-
-            const playersInfoUnfiltered = playerInfoToBeCompared[0] !== null && playerInfoToBeCompared[0] instanceof (Array) ? playerInfoToBeCompared[0].map(element => element)
-                : undefined
-
-            const playersInfo = playersInfoUnfiltered.filter(el => el.name !== '' || undefined)
-
-            if (!playersInfo) {
-                console.error(chalk.red("************* Error has occurred whilst fetching first set of new players so waiting for 30 seconds before next request to server"))
-                firstJob.cancelNext(true)
-                secondJob.cancelNext(true)
-            }
-
-            if (playersInfo.length == 0 || serverInfo[0].playersnum == 0) {
-                console.log(chalk.green("No one is on the server yet! New Players: ") + JSON.stringify(playersInfo, null, 4))
-                console.log(chalk.green("Cancelling current jobs "))
-                // TODO: Wait or not
-                firstJob.cancel(true)
-                secondJob.cancel(true)
-            }
-
-            for (i = 0; i < playersInfo.length; i++) {
-                if (!playersInfo[i] || !playersInfo[i].name) {
-                    console.error("PlayersInfo at index " + [i] + " has been skipped")
-                }
-                else {
-                    const endpointRequest = axios.post(`${endpoint}serverstats/temporaryData?name=${encodeURIComponent(playersInfo[i].name)}&score=${playersInfo[i].score}&time=${playersInfo[i].duration}&tableName=playersComparisonFirst`, {}, axiosBasicAuthConfig)
-                    playersToBePushedToTemporaryTable.push(endpointRequest)
-                }
-            }
-
-            await Promise.all(playersToBePushedToTemporaryTable);
-
-            const completedNow = moment().format('HH:mm:ss')
-            console.log(chalk.blue('Completed first job at Time: ', chalk.blueBright(completedNow)));
-        });
-
-        const secondJob = schedule.scheduleJob({ rule: '14-59/15 * * * * *' }, async (fireDate) => {
-            console.log(chalk.hex('#DEADED').bold('running second task ************************************************************************************************************************************************************'));
-            console.log('Second starting now: ' + moment().format('ss') + " but really started at " + fireDate);
-
-            const playersToBePushedToTemporaryTable = [];
-
-            const playersInfoToBeCompared = await getNewPlayers()
-                .then(eachObject => (
-                    eachObject
-                        .map(element => element.directPlayerInfo)
-                        .filter(el => el != null)))
-
-            const serverInfoUnfiltered = await getNewPlayers();
-            const serverInfo = serverInfoUnfiltered
-                .map(element => element.directQueryInfo)
-                .filter(el => el != null)
-
-            const playersInfoUnfiltered = playersInfoToBeCompared[0] !== null && playersInfoToBeCompared[0] instanceof (Array) ? playersInfoToBeCompared[0].map(element => element)
-                : undefined
-
-            const playersInfo = playersInfoUnfiltered.filter(el => el.name !== '' || undefined)
-
-            if (!playersInfo) {
-                console.error(chalk.red("************* Error has occurred whilst fetching second set of new players so waiting for 30 seconds before next request to server ------- cancelling current scheduled jobs"))
-                firstJob.cancel(true)
-                secondJob.cancel(true)
-            }
-
-            for (i = 0; i < playersInfo.length; i++) {
-                const endpointRequest = axios.post(`${endpoint}serverstats/temporaryData?name=${encodeURIComponent(playersInfo[i].name)}&score=${playersInfo[i].score}&time=${playersInfo[i].duration}&tableName=playersComparisonSecond`, {}, axiosBasicAuthConfig)
-                playersToBePushedToTemporaryTable.push(endpointRequest)
-            }
-
-            await Promise.all(playersToBePushedToTemporaryTable);
-
-            // Now that we have sent both players to the database - compare them both
-            console.log("********* START COMPARISON ***************")
-            const oldPlayersUnfiltered = await axios.get(`${endpoint}serverstats/allRows?tableName=playersComparisonFirst`, axiosBasicAuthConfig).then(element => element.data.result)
-            const newPlayersUnfiltered = await axios.get(`${endpoint}serverstats/allRows?tableName=playersComparisonSecond`, axiosBasicAuthConfig).then(element => element.data.result)
-
-            // Remove entries where they have just joined and server hasn't loaded name yet
-            const oldPlayers = oldPlayersUnfiltered.filter(el => el.name !== '' || undefined)
-            const newPlayers = newPlayersUnfiltered.filter(el => el.name !== '' || undefined)
-
-            // Compare both arrays with each other and see which elements don't exist in other one
-
-            // They have left and remove from oldPlayers array
-            if (!Array.isArray(oldPlayers) || !oldPlayers.length == 0 || serverInfo[0].playersnum >= 0) {
-                var z = oldPlayers.length
-                while (z--) {
-                    const playerHasLeft = _.findIndex(newPlayers, { name: oldPlayers[z].name }) === -1 ? true : false;
-                    if (playerHasLeft) {
-                        console.log(oldPlayers[z].name + " has abandoned the battle")
-                        const endpointRequest = axios.post(`${endpoint}serverstats/lastLogin?name=${encodeURIComponent(oldPlayers[z].name)}`, {}, axiosBasicAuthConfig)
-                        postRequests.push(endpointRequest)
-                        // remove from array
-                        const index = oldPlayers.indexOf(oldPlayers[z].name);
-                        if (index > -1) {
-                            oldPlayers.splice(index, 1);
-                        }
-                    }
-                }
-            }
-
-            // They have joined and remove from newPlayers array
-            if (!Array.isArray(newPlayers) || !newPlayers.length == 0 || serverInfo[0].playersnum >= 0) {
-                var y = newPlayers.length
-                while (y--) {
-                    const playerHasJoined = _.findIndex(oldPlayers, { name: newPlayers[y].name }) === -1 ? true : false;
-                    if (playerHasJoined) {
-                        console.log(newPlayers[y].name, " has joined the server")
-                        const endpointRequest = axios.post(`${endpoint}serverstats/?name=${encodeURIComponent(newPlayers[y].name)}`, {}, axiosBasicAuthConfig)
-                        postRequests.push(endpointRequest)
-                        // remove from array
-                        const findIndex = _.findIndex(newPlayers, { name: newPlayers[y].name })
-
-                        if (findIndex > -1) {
-                            newPlayers.splice(findIndex, 1);
-                        }
-                    }
-                }
-            }
-
-            for (i = 0; i < newPlayers.length; i++) {
-
-                let scoreDifference = 0;
-
-                newPlayerIndex = _.findIndex(newPlayers, { name: oldPlayers[i].name }) != -1 ? _.findIndex(newPlayers, { name: oldPlayers[i].name }) : _.findIndex(newPlayers, { name: newPlayers[i].name })
-                oldPlayerIndex = _.findIndex(oldPlayers, { name: newPlayers[i].name }) != -1 ? _.findIndex(oldPlayers, { name: newPlayers[i].name }) : _.findIndex(oldPlayers, { name: oldPlayers[i].name })
-
-                if (newPlayers[newPlayerIndex].score != oldPlayers[oldPlayerIndex].score) {
-                    if (newPlayers[newPlayerIndex].score < oldPlayers[oldPlayerIndex].score) {
-                        scoreDifference = oldPlayers[oldPlayerIndex].score - newPlayers[newPlayerIndex].score
-                        logger.log({
-                            level: 'info',
-                            message: `${newPlayers[newPlayerIndex].name + "'s score is less than the old one as ******** new score is " + newPlayers[newPlayerIndex].score + " and old score is " + oldPlayers[oldPlayerIndex].score + " with difference " + scoreDifference}`,
-                        });
-                        const endpointRequest = axios.post(`${endpoint}serverstats/pointsSpent?name=${encodeURIComponent(newPlayers[newPlayerIndex].name)}&pointsSpent=${scoreDifference >= 90 ? 0 : scoreDifference}`, {}, axiosBasicAuthConfig)
-                        postRequests.push(endpointRequest);
-                    }
-                    else if (newPlayers[newPlayerIndex].score > oldPlayers[oldPlayerIndex].score) {
-                        scoreDifference = newPlayers[newPlayerIndex].score - oldPlayers[oldPlayerIndex].score
-                        logger.log({
-                            level: 'info',
-                            message: `${newPlayers[newPlayerIndex].name + "'s score is more than the old one as ******** new score is " + newPlayers[newPlayerIndex].score + " and old score is " + oldPlayers[oldPlayerIndex].score + " with difference " + scoreDifference}`,
-                        });
-                        const endpointRequest = axios.post(`${endpoint}serverstats/kills?name=${encodeURIComponent(newPlayers[newPlayerIndex].name)}&kills=${scoreDifference % 2 == 0 ? scoreDifference / 2 : scoreDifference}`, {}, axiosBasicAuthConfig)
-                        postRequests.push(endpointRequest)
-                    }
-                }
-                else {
-                    logger.log({
-                        level: 'info',
-                        message: `${newPlayers[newPlayerIndex].name + "'s score hasn't changed ******** because new score is " + newPlayers[newPlayerIndex].score + " and old score is " + oldPlayers[oldPlayerIndex].score}`,
-                    });
-                    const endpointRequest = axios.post(`${endpoint}serverstats/?name=${encodeURIComponent(newPlayers[newPlayerIndex].name)}`, {}, axiosBasicAuthConfig)
-                    postRequests.push(endpointRequest)
-                }
-            }
-
-            // Send all endpoint requests from above
-            Promise.all(postRequests)
-                .then(console.log(chalk.whiteBright("Sent remaining players to database")))
-                .catch(console.error)
-
-            //  Need to truncate temp tables playersComparisonFirst and playersComparisonSecond
-            pool.getConnection((err, connection) => {
-                if (err) console.log(err);
-                connection.query(`TRUNCATE playersComparisonFirst;`, (err, result, fields) => {
-                    if (err) console.log(err);
-                    if (result) {
-                        console.log("Reset all rows in playersComparisonFirst table")
-                    }
-                    if (fields) console.log(fields);
-                });
-                connection.query(`TRUNCATE playersComparisonSecond;`, (err, result, fields) => {
-                    if (err) console.log(err);
-                    if (result) {
-                        console.log("Reset all rows in playersComparisonSecond table")
-                    }
-                    if (fields) console.log(fields);
-                });
-                connection.release();
-            });
-
-            const completedNow = moment().format('HH:mm:ss')
-            console.log(chalk.blue('Completed second job at Time: ', chalk.blueBright(completedNow)));
-
-        });
-
-        // Do not do the first scheduled run of secondJob as firstJob needs to be in front with a 15 second offset
-        secondJob.cancelNext(true)
-
-    }
-
-    else {
-        console.log("Already running, restart endpoint/dyno in heroku and call repeatedRequests again")
-        return res.status(404).json({
-            error: {
-                message: "This endpoint has already been called and was called at: " + timestamp
-            }
-        });
-    }
 })
 
 module.exports = router
